requirements.txt
Django
djangorestframework
psycopg2
python-dotenv
djangorestframework-simplejwt
drf-yasg
Pillow>=9.0.0
